Feature: Se agregó un objeto de respuesta generica para los controladores de las apis

+++ src/main/java/com/mysql/demo/api/users/repository/User.java
@@ -9,8 +9,8 @@ import java.time.LocalDate;
 @Table(name = "users")
 public class User {
     @Id
-    @GeneratedValue(strategy=GenerationType.AUTO)
-    private Integer id;
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    private Long id;
     @Column(name = "username")
     private String username;
     @Column(name = "email")
@@ -24,11 +24,11 @@ public class User {
     public User() {
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
